requirements.txt
# GarbageSort AI — Requirements
# ================================
# Edge-device optimized waste classification system
# Install: pip install -r requirements.txt

# ── Core ML ───────────────────────────────────────────────────────────────────
tensorflow>=2.13.0              # Model training & inference
numpy>=1.26.0
scikit-learn>=1.3.0

# ── Vision ────────────────────────────────────────────────────────────────────
opencv-python>=4.9.0            # Live camera capture
Pillow>=10.0.0                  # Image handling

# ── Web Dashboard ─────────────────────────────────────────────────────────────
streamlit>=1.32.0
plotly>=5.20.0                  # Interactive charts
pandas>=2.2.0

# ── Report Generation ─────────────────────────────────────────────────────────
fpdf2>=2.7.0                    # PDF report export
qrcode[pil]>=7.4.0              # QR codes for recycling guides

# ── System Monitoring ─────────────────────────────────────────────────────────
psutil>=5.9.0                   # CPU / RAM stats for edge info page

# ── Visualization ─────────────────────────────────────────────────────────────
matplotlib>=3.8.0
seaborn>=0.13.0
